interface implementation

+++ src/main/java/com/laura/carpaciu/controllers/work/WorkController.java
@@ -90,7 +90,7 @@ public class WorkController {
 	}
 
 	@GetMapping("/delete")
-	public String deleteWork(@RequestParam("workId") Long id) {
+	public String deleteWork(@RequestParam("workId") int id) {
 
 		workService.deleteWork(id);
